+ hacked printf format errors for XML_Size in mootExpatParser.cc: use '%lu' format and static_cast<>
git-svn-id: file:///home/svn/dev/moot/trunk@1630 4a515de8-beb3-11dc-84d6-2b654028afac

+++ src/libmoot/mootExpatParser.cc
@@ -325,8 +325,8 @@ void mootExpatParser::xpcarp(const char *fmt, ...)
   fprintf(stderr, " in `%s' at line %lu, column %lu: %s\n",
 	  (xp_istream && !xp_istream->name.empty()
 	   ? xp_istream->name.c_str() : "(unknown)"),
-	  XML_GetCurrentLineNumber(parser),
-	  XML_GetCurrentColumnNumber(parser),
+	  static_cast<long unsigned int>(XML_GetCurrentLineNumber(parser)),
+	  static_cast<long unsigned int>(XML_GetCurrentColumnNumber(parser)),
 	  XML_ErrorString(XML_GetErrorCode(parser)));
 }
